% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolgraph-package.r
\docType{package}
\name{coolgraph}
\alias{coolgraph}
\alias{coolgraph-package}
\title{coolgraph. making complex graphs easy}
\description{
This package has been inspired by my research group and there desire to make awesome graphs without really knowing much about R. In this package I will be introducing three new funcitons, one that allows you to make a simple heatmap, another that lets you copare 2 continous variables and facet over anothe 2 categorical variables, and a 3rd that allows you to compare 3 continous variables by mapping the 3rd variable to the size of the data points. It contains 3 major functions, heatmap(), facet2D(), and scalesize().
}

