% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizescale.R
\name{sizescale}
\alias{sizescale}
\title{sizescale}
\usage{
sizescale(data, x1, y1, scale, xlab = "x-axis", ylab = "y-axis",
  legendlab = "legend")
}
\arguments{
\item{data}{is a data frame}

\item{x1}{is a column of a data frame of the form data$variable name, containing int, dbl, or numeric data}

\item{y1}{is a column of a data frame of the form data$variable name, containing int, dbl, or numeric data}

\item{scale}{is a column of a data frame of the form data$variable name, containing int, dbl, or numeric data}

\item{xlab}{is the x-axis label - must be character data.}

\item{ylab}{is the y-axis label - must be character data.}

\item{legendlab}{is the legend label - must be character data.}
}
\description{
This function is directly inspired by my lab's request to plot 3 continous variables on one graph, as such this function plots one variable on the x axis, one variable on the y axis, and maps on variable to the size of each point, which is drawn as an outline to reduce cluttering of the data.
}
\examples{
sizescale(mtcars, mtcars$wt, mtcars$mpg, mtcars$hp, "wt", "mpg", "hp")
sizescale(mtcars, mtcars$wt, mtcars$mpg, mtcars$hp, "weight", "mpg", "horse power")
sizescale(mtcars, mtcars$disp, mtcars$mpg, mtcars$hp, "size", "mpg", "horse power")
sizescale(mtcars, mtcars$disp, mtcars$mpg, mtcars$hp)
}

